Simplify and fix percent encoding in URL.addParameter (#78)
When adding a new query parameter to an existing URL, manually
percent encode the name and value and add them to
percentEncodedQueryItems. When percent-encoding the parameter, use a
modified allowed characters set that does not include "URI query
reserved characters" as they need to be percent encoded (when part
of the parameter name or value).