Conditionally allow reserved characters in URL.addParameter (#80)
Task/Issue URL: https://app.asana.com/0/1163321984198618/1202033667108435/f

Description:
Add an optional argument to URL.addParameter to specify reserved characters
(as per RFC 3986) that should not be percent-encoded, for example because 
they play a special role in the payload, as defined by the API provider.

By default we percent-encode all reserved characters, which should be a safe default,
unless specific API providers instruct to allow certain characters. This is the case for
Website Breakage reports where comma is used to delimit blocked trackers domains.

This is an macOS-only change, as iOS does not use the URL.addParameter at this time.
Furthermore, it's an opt-in API, so by default nothing will change for existing callers.