Email Manager keychain error reporting (#123)
Required:

Task/Issue URL: https://app.asana.com/0/414235014887631/1202679436624547/f
iOS PR: duckduckgo/iOS#1258
macOS PR: more-duckduckgo-org/macos-browser#662
What kind of version bump will this require?: Patch

Description:

This PR adds error reporting to the email manager's keychain access. It does this by adding a new delegate method to the email manager's request delegate, which then requests that the client codebase fire a pixel on BSK's behalf.

Errors are reported when data from the keychain cannot be read correctly, or if the keychain itself has an error, unless the error is errSecItemNotFound in which case the usual nil value is returned.

Note: This functionality was deliberately only added to non-waitlist functions.

Extra Note: There are SwiftLint failures that exist in the modified files that have been around for a long time, they're not worth fixing as a part of these changes.