Send the email cohort in user info when signing in and out. (#197)
Required:

Task/Issue URL: https://app.asana.com/0/414235014887631/1203393655574808/f
iOS PR: duckduckgo/iOS#1418
macOS PR: more-duckduckgo-org/macos-browser#881
What kind of version bump will this require?: Minor

Description:

This PR updates the notifications sent by the email manager when signing in and out to include the cohort as a part of the user info. This is done as the cohort is deleted before the sign out notification is sent, so you weren't able to access it when the user is signing out.