Extract SecureVault code into SecureStorage module (#430)
Task/Issue URL: https://app.asana.com/0/0/1205119749390492/f
Tech Design URL:
CC: @Bunn @jotaemepereira

Description:

This PR extracts the reusable SecureVault logic out into a new SecureStorage module.

This module is used by Autofill, and is usable by the client apps directly to implement secure storage on their own. The API for this new module matches the existing API as much as possible, and the implementation of the key/crypto providers is made easier by providing default implementations of the generic functions.

Note that not all areas of the code have been renamed; for example, there is still a struct called SecureVaultModels which I would like to rename AutofillModels, but in this case this symbol is referenced over 600 times, so I plan to do this in a follow-up renaming PR to keep this PR as small as possible (it's already quite large).
# Conflicts:
#	Sources/BrowserServicesKit/SecureVault/SecureVaultManager.swift
#	Tests/BrowserServicesKitTests/SecureVault/SecureVaultManagerTests.swift