Improve handling lists in Sync (#575)
Task/Issue URL: https://app.asana.com/0/0/1205884549503153/f

Description:
Improve handling of lists in syncable models' payloads (which currently is limited to bookmarks
folders and favorites folders children). This change updates how lists are sent to Sync in PATCH
payloads by including current state as well as the set of removed and inserted items.
To allow for this, last sync payload is kept for all bookmarks folders in
lastChildrenPayloadReceivedFromSync property added on BookmarkEntity entity.
If the payload is not available (e.g. first sync), the legacy mechanism is used (i.e. sending only
the current contents of the list). The schema of a Sync response payload does not change so
the implementation handles it too.