autofill: send current language in runtime config
The HTML autofill UI will be translated shortly, but doing so is only
possible once that UI knows which language to use. Send the current
locale's two-character language code as part of the privacy config (aka
'runtime config' in the autofill codebase).

This is a backwards-compatible change, since 'language' is currently an
optional input to the autofill HTML pages.