autofill: send current language in runtime config (#808)
Task/Issue URL: https://app.asana.com/0/11984721910118/1206933877192910/f
iOS PR: N/A
macOS PR: N/A
What kind of version bump will this require?: Patch. This is backwards-compatible, and the autofill codebase falls back to 'en' when it doesn't receive a language.

Description:
The HTML autofill UI will be translated shortly, but doing so is only possible once that UI knows which language to use. Send the current locale's two-character language code as part of the privacy config (aka 'runtime config' in the autofill codebase).

This is a backwards-compatible change, since 'language' is currently an optional input to the autofill HTML pages.