Refactor Core Data usage in RemoteMessagingStore (#1013)
Task/Issue URL: https://app.asana.com/0/1201048563534612/1208493869353425/f

Description:
This change fixes Core Data usage in RemoteMessagingStore. Two types of managed object
contexts are now in use. One of them is long-lived and is used for saving data to the Store.
The other is created on demand and used for fetching data. The long-lived context uses asynchronous
Core Data API to avoid potential deadlocks, and is wrapped in async functions for simpler use.
Single instance of a long-lived context helps prevent merge conflicts.